+++ lib/app/modules/starred_messages/controllers/starred_messages_controller.dart
@@ -13,8 +13,6 @@ import 'package:mirrorfly_uikit_plugin/app/data/helper.dart';
 import 'package:share_plus/share_plus.dart';
 
 import '../../../common/constants.dart';
-import '../../chat/views/chat_view.dart';
-import '../../chat/views/forwardchat_view.dart';
 
 class StarredMessagesController extends FullLifeCycleController with FullLifeCycleMixin {
   var starredChatList = List<ChatMessageModel>.empty(growable: true).obs;
@@ -251,9 +249,9 @@ class StarredMessagesController extends FullLifeCycleController with FullLifeCyc
   checkBusyStatusForForward(BuildContext context) async {
     var busyStatus = await Mirrorfly.isBusyStatusEnabled();
     if (!busyStatus.checkNull()) {
-      forwardMessage(context);
+      if(context.mounted)forwardMessage(context);
     } else {
-      showBusyStatusAlert(forwardMessage, context);
+      if(context.mounted)showBusyStatusAlert(forwardMessage, context);
     }
   }
 
@@ -331,7 +329,7 @@ class StarredMessagesController extends FullLifeCycleController with FullLifeCyc
 
   copyTextMessages() {
     Clipboard.setData(
-        ClipboardData(text: selectedChatList[0].messageTextContent));
+        ClipboardData(text: selectedChatList[0].messageTextContent.toString()));
     clearChatSelection(selectedChatList[0]);
     toToast("1 Text Copied Successfully to the clipboard");
   }